# backend/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Contrôle Parental API'
  description: "API REST pour l'application de Contrôle Parental - gestion des familles, enfants, appareils, règles de filtrage, temps d'écran, alertes et rapports."
  version: 1.0.0
servers:
  - url: 'http://localhost:8000'
    description: 'Serveur local'
paths:
...
  /api/notifications:
    get:
      tags:
        - Notifications
      summary: 'Liste des notifications'
      description: "Retourne les notifications de l'utilisateur authentifié (paginées)"
      operationId: notificationsIndex
      parameters:
        -
          name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        200:
          description: 'Liste paginée des notifications'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Liste des notifications'
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  meta:
                    type: object
                  links:
                    type: object
                type: object
        401:
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          sanctum: []
  '/api/notifications/{notification}':
    get:
      tags:
        - Notifications
      summary: "Détails d'une notification"
      operationId: notificationsShow
      parameters:
        -
          name: notification
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Détails de la notification'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Détails de la notification'
                  data:
                    $ref: '#/components/schemas/Notification'
                type: object
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
      security:
        -
          sanctum: []
  '/api/notifications/{notification}/read':
    patch:
      tags:
        - Notifications
      summary: 'Marquer une notification comme lue'
      operationId: notificationsMarkAsRead
      parameters:
        -
          name: notification
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Notification marquée comme lue'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Notification marquée comme lue'
                  data:
                    $ref: '#/components/schemas/Notification'
                type: object
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
      security:
        -
          sanctum: []
  /api/notifications/mark-all-read:
    post:
      tags:
        - Notifications
      summary: 'Marquer toutes les notifications comme lues'
      operationId: notificationsMarkAllAsRead
      responses:
        200:
          $ref: '#/components/responses/Success'
        401:
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          sanctum: []
...
  /api/reports:
    get:
      tags:
        - Reports
      summary: 'Liste des rapports'
      description: 'Retourne les rapports (paginés), filtrable par child_id et period_type'
      operationId: reportsIndex
      parameters:
        -
          name: page
          in: query
          schema:
            type: integer
            default: 1
        -
          name: child_id
          in: query
          schema:
            type: integer
        -
          name: period_type
          in: query
          schema:
            type: string
      responses:
        200:
          description: 'Liste paginée des rapports'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Liste des rapports'
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Report'
                  meta:
                    type: object
                  links:
                    type: object
                type: object
        401:
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          sanctum: []
  '/api/reports/{report}':
    get:
      tags:
        - Reports
      summary: "Détails d'un rapport"
      operationId: reportsShow
      parameters:
        -
          name: report
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Détails du rapport'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Détails du rapport'
                  data:
                    $ref: '#/components/schemas/Report'
                type: object
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
      security:
        -
          sanctum: []
components:
  schemas:
...
    Notification:
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        alert_id:
          type: integer
          nullable: true
        channel:
          type: string
          example: push
        title:
          type: string
          example: 'Nouvelle alerte'
        body:
          type: string
          example: "Lucas a tenté d'accéder à..."
        status:
          type: string
          example: sent
        sent_at:
          type: string
          format: date-time
          nullable: true
        read_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
      type: object
    Report:
      properties:
        id:
          type: integer
          example: 1
        child_id:
          type: integer
          example: 1
        period_type:
          type: string
          example: weekly
        period_start:
          type: string
          format: date
        period_end:
          type: string
          format: date
        digital_health_score:
          type: integer
          example: 85
        statistics:
          type: object
        file_path:
          type: string
          nullable: true
        generated_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
      type: object
...
  responses:
    Success:
      description: 'Opération réussie'
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: true
              message:
                type: string
                example: 'Opération réussie'
              data:
                type: object
                nullable: true
            type: object
    Unauthorized:
      description: 'Non authentifié'
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Unauthenticated.
            type: object
...
    NotFound:
      description: 'Ressource introuvable'
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 'Ressource introuvable'
            type: object
...
  securitySchemes:
    sanctum:
      type: apiKey
      description: 'Enter token in format (Bearer <token>)'
      name: Authorization
      in: header
...
security:
  -
    sanctum: []